% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGeneData.R
\docType{methods}
\name{getExpressionDatasetDirectory,TrenaGeneData-method}
\alias{getExpressionDatasetDirectory,TrenaGeneData-method}
\title{local directory where expression data RData files are foudn}
\usage{
\S4method{getExpressionDatasetDirectory}{TrenaGeneData}(obj)
}
\arguments{
\item{obj}{a TrenaGeneData (or subclass) object}
}
\value{
character vector, the full path to the direcotyr
}
\description{
local directory where expression data RData files are foudn
}
