% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGeneData.R
\docType{methods}
\name{getExpressionDatasetNames,TrenaGeneData-method}
\alias{getExpressionDatasetNames,TrenaGeneData-method}
\title{one or more descriptive filenames, .RData suffix stripped off}
\usage{
\S4method{getExpressionDatasetNames}{TrenaGeneData}(obj)
}
\arguments{
\item{obj}{a TrenaGeneData (or subclass) object}
}
\value{
character vector, a list of one or more expression data sets
}
\description{
one or more descriptive filenames, .RData suffix stripped off
}
