% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGeneData.R
\docType{methods}
\name{getFootprintDatabaseNames,TrenaGeneData-method}
\alias{getFootprintDatabaseNames,TrenaGeneData-method}
\title{one or more postgress database names, assumed for now to be on khaleesi}
\usage{
\S4method{getFootprintDatabaseNames}{TrenaGeneData}(obj)
}
\arguments{
\item{obj}{a TrenaGeneData (or subclass) object}
}
\value{
character vector, a list of one or more postgres database names
}
\description{
one or more postgress database names, assumed for now to be on khaleesi
}
