% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGeneData.R
\docType{methods}
\name{getGeneSymbol,TrenaGeneData-method}
\alias{getGeneSymbol,TrenaGeneData-method}
\title{the HUGO gene symbol}
\usage{
\S4method{getGeneSymbol}{TrenaGeneData}(obj)
}
\arguments{
\item{obj}{a TrenaGeneData (or subclass) object}
}
\value{
character, the HUGO symbol for this gene
}
\description{
the HUGO gene symbol
}
