% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGeneData.R
\docType{methods}
\name{getGenomeName,TrenaGeneData-method}
\alias{getGenomeName,TrenaGeneData-method}
\title{for instance, hg38}
\usage{
\S4method{getGenomeName}{TrenaGeneData}(obj)
}
\arguments{
\item{obj}{a TrenaGeneData (or subclass) object}
}
\value{
character, the standard code for the genome build for the data offered here
}
\description{
for instance, hg38
}
