% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGeneData.R
\docType{class}
\name{TrenaGeneData-class}
\alias{TrenaGeneData-class}
\alias{.TrenaGeneData}
\alias{TrenaGeneData}
\title{Define an object of class TrenaGeneData}
\usage{
TrenaGeneData(geneSymbol, genomeName,
  footprintDatabaseNames = NA_character_,
  expressionDatasetDirectory = NA_character_,
  expressionDatasetNames = NA_character_)
}
\arguments{
\item{geneSymbol}{character, the HUGO gene name}

\item{footprintDatabaseNames, }{character vector, zero or more Postgres databases}

\item{expressionDatasetNames, }{character vector, zero or more RData filenames}
}
\value{
An object of the TrenaGeneData class
}
\description{
an abstract base class
}
